kube/second/front-back-mysql/backend/requirements.txt
Flask==2.0.0
mysql-connector-python==8.0.26
Werkzeug==2.0.0
